моки переместила в фикстуры, без примитивных фикстур создаю в тестах объекты, убрала helpers.py
